2. Фрагменты приложения, структура MVVM, панель навигации
