+++ app/src/main/java/com/technocopy/weatherapp/view/CurrentWeatherFragment.kt
@@ -1,0 +1,33 @@
+package com.technocopy.weatherapp.view
+
+import androidx.lifecycle.ViewModelProvider
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.technocopy.weatherapp.viewmodel.CurrentWeatherViewModel
+import com.technocopy.weatherapp.R
+
+class CurrentWeatherFragment : Fragment() {
+
+    companion object {
+        fun newInstance() = CurrentWeatherFragment()
+    }
+
+    private lateinit var viewModel: CurrentWeatherViewModel
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.current_weather_fragment, container, false)
+    }
+
+    override fun onActivityCreated(savedInstanceState: Bundle?) {
+        super.onActivityCreated(savedInstanceState)
+        viewModel = ViewModelProvider(this).get(CurrentWeatherViewModel::class.java)
+        // TODO: Use the ViewModel
+    }
+
+}
