+++ app/src/main/java/com/technocopy/weatherapp/view/MainActivity.kt
@@ -1,11 +1,8 @@
-package com.technocopy.weatherapp
+package com.technocopy.weatherapp.view
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.view.View
-import android.widget.Button
-import android.widget.TextView
-import android.widget.Toast
+import com.technocopy.weatherapp.R
 
 class MainActivity : AppCompatActivity() {
