+++ app/src/main/java/com/technocopy/weatherapp/view/TenDaysWeatherFragment.kt
@@ -1,4 +1,4 @@
-package com.technocopy.weatherapp
+package com.technocopy.weatherapp.view
 
 import androidx.lifecycle.ViewModelProvider
 import android.os.Bundle
@@ -6,6 +6,8 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.technocopy.weatherapp.R
+import com.technocopy.weatherapp.viewmodel.TenDaysWeatherViewModel
 
 class TenDaysWeatherFragment : Fragment() {
