+++ app/src/main/java/com/technocopy/weatherapp/viewmodel/SettingsViewModel.kt
@@ -1,0 +1,7 @@
+package com.technocopy.weatherapp.viewmodel
+
+import androidx.lifecycle.ViewModel
+
+class SettingsViewModel : ViewModel() {
+    // TODO: Implement the ViewModel
+}
