+++ app/src/main/res/layout/settings_fragment.xml
@@ -3,12 +3,12 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.CurrentWeatherFragment">
+    tools:context=".view.SettingsFragment">
 
     <TextView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:gravity="center"
-        android:text="@string/current_weather" />
+        android:text="@string/settings" />
 
 </FrameLayout>
