+++ app/src/main/res/layout/ten_days_weather_fragment.xml
@@ -3,11 +3,12 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".TenDaysWeatherFragment">
+    tools:context=".view.TenDaysWeatherFragment">
 
     <TextView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:text="Hello" />
+        android:gravity="center"
+        android:text="@string/ten_days" />
 
 </FrameLayout>
