+++ app/src/main/res/navigation/main_navigation.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/main_navigation"
+    app:startDestination="@id/currentWeatherFragment4">
+    <fragment
+        android:id="@+id/currentWeatherFragment4"
+        android:name="com.technocopy.weatherapp.view.CurrentWeatherFragment"
+        android:label="current_weather_fragment"
+        tools:layout="@layout/current_weather_fragment" />
+    <fragment
+        android:id="@+id/tenDaysWeatherFragment2"
+        android:name="com.technocopy.weatherapp.view.TenDaysWeatherFragment"
+        android:label="ten_days_weather_fragment"
+        tools:layout="@layout/ten_days_weather_fragment" />
+    <fragment
+        android:id="@+id/settingsFragment3"
+        android:name="com.technocopy.weatherapp.view.SettingsFragment"
+        android:label="settings_fragment"
+        tools:layout="@layout/settings_fragment" />
+</navigation>
