+++ app/src/main/res/values/strings.xml
@@ -9,4 +9,5 @@
     <string name="_10_days">10 days</string>
     <string name="settings">Settings</string>
     <string name="current_weather">Current weather</string>
+    <string name="ten_days">Ten days forecast</string>
 </resources>
